@@ -23,20 +23,20 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.1.0",
-    "@eslint/js": "^9.12.0",
+    "@eslint/js": "^9.13.0",
     "@picocss/pico": "^2.0.6",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
-    "@testing-library/jest-dom": "^6.5.0",
-    "@testing-library/svelte": "^5.2.3",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
+    "@testing-library/jest-dom": "^6.6.2",
+    "@testing-library/svelte": "^5.2.4",
     "@tsconfig/svelte": "^5.0.3",
-    "@typescript-eslint/eslint-plugin": "^8.9.0",
+    "@typescript-eslint/eslint-plugin": "^8.12.0",
     "@vite-pwa/assets-generator": "^0.2.6",
-    "@vitest/coverage-v8": "^2.1.3",
-    "@vitest/ui": "^2.1.3",
+    "@vitest/coverage-v8": "^2.1.4",
+    "@vitest/ui": "^2.1.4",
     "autoprefixer": "^10.4.20",
-    "eslint": "^9.12.0",
+    "eslint": "^9.13.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-svelte": "^2.44.1",
+    "eslint-plugin-svelte": "^2.46.0",
     "globals": "^15.11.0",
     "husky": "^9.1.6",
     "jsdom": "^25.0.1",
@@ -44,19 +44,19 @@
     "prettier": "^3.3.3",
     "prettier-plugin-css-order": "^2.1.0",
     "prettier-plugin-svelte": "^3.2.7",
-    "sass": "^1.79.5",
+    "sass": "^1.80.4",
     "stylelint": "^16.10.0",
     "stylelint-config-standard-scss": "^13.0.0",
     "stylelint-prettier": "^5.0.2",
-    "svelte": "^4.2.19",
+    "svelte": "^5.1.3",
     "svelte-check": "^4.0.5",
-    "svelte-i18n": "^4.0.0",
+    "svelte-i18n": "^4.0.1",
     "svelte-preprocess": "^6.0.3",
-    "tslib": "^2.7.0",
+    "tslib": "^2.8.0",
     "typescript": "^5.6.3",
-    "vite": "^5.4.9",
+    "vite": "^5.4.10",
     "vite-plugin-mkcert": "^1.17.6",
     "vite-plugin-pwa": "^0.20.5",
-    "vitest": "^2.1.3"
+    "vitest": "^2.1.4"
   }
 }
