@@ -1,10 +1,10 @@
 <script lang="ts">
-  let count: number = 0;
+  let count: number = $state(0);
   const increment = () => {
     count += 1;
   };
 </script>
 
-<button on:click="{increment}">
+<button onclick={increment}>
   count is {count}
 </button>
