@@ -2,8 +2,9 @@ import '@picocss/pico';
 import './i18n.js';
 import './app.css';
 import App from './App.svelte';
+import { mount } from "svelte";
 
-const app = new App({
+const app = mount(App, {
   target: document.getElementById('app')!,
 });
 
