@@ -1,13 +1,14 @@
-import { defineConfig } from 'vite'
-import { svelte } from '@sveltejs/vite-plugin-svelte'
+import { defineConfig } from 'vite';
+import { svelte } from '@sveltejs/vite-plugin-svelte';
 
-export default defineConfig({
-    plugins: [
-        svelte({ hot: !process.env.VITEST }),
-    ],
-    test: {
-        globals: true,
-        environment: 'jsdom',
-        setupFiles: ['./vitest-setup.ts'],
-    },
-})
+export default defineConfig(({ mode }) => ({
+  test: {
+    globals: true,
+    environment: 'jsdom',
+    setupFiles: ['./vitest-setup.ts'],
+  },
+  resolve: {
+    conditions: mode === 'test' ? ['browser'] : [],
+  },
+  plugins: [svelte({ hot: !process.env.VITEST })],
+}));
